requirements.txt
flask
tensorflow
pandas
numpy
scikit-learn
pillow
tqdm
